+++ Engine/source/engin.cpp
@@ -12,15 +12,22 @@
 //static SDL_Window* _window = NULL;
 
 using namespace homer;
+/// <summary>
+/// Function init where everything is initialise
+/// </summary>
+/// <param name="name"></param>
+/// <param name="w"></param>
+/// <param name="h"></param>
+/// <returns></returns>
 bool Engin::Init(const char* name, int w, int h) {
-
+	
 	m_Input = new SdlInput();
 	m_IsInit = true;
 	m_Graphics = new SDLGraphics();
 	m_Audio = new SDLAudio();
 	m_Graphics->Initialize(name, w, h);
 
-	// la music ne marche pas.
+	// la music ne marche pas?
 	//m_Audio->PlayMusic(m_Audio->LoadMusic("assets/Coin.wav"));
 
 #ifdef _DEBUG
@@ -32,16 +39,20 @@ bool Engin::Init(const char* name, int w, int h) {
 #endif
 	return true;
 }
-
+/// <summary>
+/// Function Start with the delta time and the 60 fps
+/// </summary>
+/// <param name=""></param>
 void Engin::Start(void) {
 	if (!m_IsInit) {
 		if (!Init("Unknow title",800,600)) {
 			return;
 		}
 	}
+	
 	m_IsRunning = true;
 	clock_t _end = clock();
-
+	
 	while (m_IsRunning) {
 		const clock_t _start = clock();
 		float _dt = (_start - _end) * 0.001f;
@@ -52,14 +63,18 @@ void Engin::Start(void) {
 		float elapse = (_end - _start) * 16.666f * 0.001f;
 		if(elapse > 0)
 		{
-			Sleep(elapse);
+			Sleep(static_cast<DWORD>(elapse));
 		}
 		_end = _start;
 		
 	}
+	
 	ShutDown();
 }
-
+/// <summary>
+/// Where the input are call
+/// </summary>
+/// <param name=""></param>
 void Engin::ProcessInput(void)
 {
 	m_Input->Update();
@@ -68,6 +83,10 @@ void Engin::ProcessInput(void)
 static float x = 0.0f;
 static float y = 0.0f;
 static float speed = 100;
+/// <summary>
+/// Where everuthing is updated
+/// </summary>
+/// <param name="dt"></param>
 void Engin::Update(float dt)
 {
 	if (m_Input->IsKeyDown(7)) {
@@ -98,7 +117,10 @@ void Engin::Update(float dt)
 #endif
 	
 }
-
+/// <summary>
+/// Render the texture,string and the rectangle
+/// </summary>
+/// <param name=""></param>
 void Engin::Render(void)
 {
 	m_Graphics->Clear();
@@ -110,19 +132,24 @@ void Engin::Render(void)
 
 	m_Graphics->DrawTexture(m_Graphics->LoadTexture("assets/Candle.png"), _carre,Color::Red);
 	m_Graphics->DrawString("allo", m_Graphics->LoadFont("assets/Happy.ttf",50),60,60,Color::Blue);
-	m_Graphics->DrawRect(x, y, 100, 100, Color::Red);
+	m_Graphics->DrawRect(static_cast<int>(x), static_cast<int>(y), 100, 100, Color::Red);
 
 	m_Graphics->Present();
 }
-
+/// <summary>
+/// Close the engine
+/// </summary>
 void Engin::Quit()
 {
 	m_IsRunning = false;
 }
-
+/// <summary>
+/// Shutdown everything
+/// </summary>
+/// <param name=""></param>
 void Engin::ShutDown(void)
 {
 	m_IsRunning = false;
-	//SDL_DestroyRenderer(_renderer);
+	//SDL_DestroyRenderer(m_Renderer);
 	SDL_Quit();
 }
